requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3

# Facebook Ads
facebook-business==20.0.0

# Shopify
ShopifyAPI==12.6.0

# Anthropic
anthropic==0.40.0

# Media processing (adicionar depois quando for implementar análise de vídeo)
# openai-whisper==20231117
# ffmpeg-python==0.2.0

# HTTP client
httpx==0.27.2
requests==2.32.3

# Async support
anyio==4.6.2

# Environment
python-dotenv==1.0.1

# Utilities
tenacity==9.0.0
python-dateutil==2.9.0
